Rework buildNHeaders to be faster, move it to ChainUnitTest
